Untangle test-suites and jvm plugin projects
Move JvmTestSuite types to plugins-jvm-test-suite by creating a JvmTestSuitePluginHelper to remove the need for these types in plugins-java-base.

This type now holds the name of the default test suite.

Also finally delete jvm test suite base project.
