+++ platforms/jvm/war/build.gradle.kts
@@ -35,6 +35,7 @@ dependencies {
     implementation(project(":plugins"))
     implementation(project(":plugins-java"))
     implementation(project(":plugins-java-base"))
+    implementation(project(":plugins-jvm-test-suite"))
     implementation(project(":testing-base"))
 
     implementation(libs.groovy)
