+++ platforms/software/build-init/src/main/java/org/gradle/buildinit/plugins/internal/LanguageLibraryProjectInitDescriptor.java
@@ -16,7 +16,7 @@
 
 package org.gradle.buildinit.plugins.internal;
 
-import org.gradle.api.plugins.JvmTestSuitePlugin;
+import org.gradle.api.plugins.internal.JvmTestSuitePluginHelper;
 import org.gradle.buildinit.plugins.internal.modifiers.BuildInitTestFramework;
 import org.gradle.buildinit.plugins.internal.modifiers.ModularizationOption;
 
@@ -44,7 +44,7 @@ public abstract class LanguageLibraryProjectInitDescriptor implements LanguageSp
     }
 
     protected BuildScriptBuilder.SuiteSpec configureDefaultTestSuite(BuildScriptBuilder buildScriptBuilder, BuildInitTestFramework testFramework, TemplateLibraryVersionProvider libraryVersionProvider) {
-        return addTestSuite(JvmTestSuitePlugin.DEFAULT_TEST_SUITE_NAME, buildScriptBuilder, testFramework, libraryVersionProvider);
+        return addTestSuite(JvmTestSuitePluginHelper.DEFAULT_TEST_SUITE_NAME, buildScriptBuilder, testFramework, libraryVersionProvider);
     }
 
     protected BuildScriptBuilder.SuiteSpec addTestSuite(String name, BuildScriptBuilder buildScriptBuilder, BuildInitTestFramework testFramework, TemplateLibraryVersionProvider libraryVersionProvider) {
