+++ subprojects/plugin-development/build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
     implementation(project(":plugins-groovy"))
     implementation(project(":plugins-java"))
     implementation(project(":plugins-java-base"))
+    implementation(project(":plugins-jvm-test-suite"))
     implementation(project(":language-java"))
     implementation(project(":plugin-use"))
     implementation(project(":publish"))
