+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/JavaTestFixturesPlugin.java
@@ -23,6 +23,7 @@ import org.gradle.api.artifacts.dsl.DependencyHandler;
 import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.plugins.internal.JavaConfigurationVariantMapping;
 import org.gradle.api.plugins.internal.JavaPluginHelper;
+import org.gradle.api.plugins.internal.JvmTestSuitePluginHelper;
 import org.gradle.api.plugins.jvm.internal.DefaultJvmFeature;
 import org.gradle.api.plugins.jvm.internal.JvmFeatureInternal;
 import org.gradle.api.tasks.SourceSet;
@@ -86,7 +87,7 @@ public abstract class JavaTestFixturesPlugin implements Plugin<Project> {
         feature.getApiConfiguration().getDependencies().add(dependencies.create(project));
 
         // The tests depend on the test fixtures.
-        SourceSet testSourceSet = JavaPluginHelper.getDefaultTestSuite(project).getSources();
+        SourceSet testSourceSet = JvmTestSuitePluginHelper.getDefaultTestSuite(project).getSources();
         Configuration testImplementation = project.getConfigurations().getByName(testSourceSet.getImplementationConfigurationName());
         testImplementation.getDependencies().add(dependencies.testFixtures(dependencies.create(project)));
